Split test_getfeature.py into multiple files
