+++ tests/gisserver/views/test_getfeature_resourceid.py
@@ -1,0 +1,117 @@
+import pytest
+
+from tests.constants import NAMESPACES
+from tests.utils import WFS_20_XSD, read_response, validate_xsd
+
+# enable for all tests in this file
+pytestmark = [pytest.mark.urls("tests.test_gisserver.urls")]
+
+
+@pytest.mark.django_db
+class TestGetFeature:
+    """All tests for the GetFeature method.
+    The methods need to have at least one datatype, otherwise not all content is rendered.
+    """
+
+    def test_resource_id(self, client, restaurant, bad_restaurant):
+        """Prove that fetching objects by ID works."""
+        response = client.get(
+            "/v1/wfs/?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0"
+            f"&RESOURCEID=restaurant.{restaurant.id}"
+        )
+        content = read_response(response)
+        assert response["content-type"] == "text/xml; charset=utf-8", content
+        assert response.status_code == 200, content
+        assert "</wfs:FeatureCollection>" in content
+
+        # Validate against the WFS 2.0 XSD
+        xml_doc = validate_xsd(content, WFS_20_XSD)
+        assert xml_doc.attrib["numberMatched"] == "1"
+        assert xml_doc.attrib["numberReturned"] == "1"
+
+        # Test sort ordering.
+        restaurants = xml_doc.findall("wfs:member/app:restaurant", namespaces=NAMESPACES)
+        names = [res.find("app:name", namespaces=NAMESPACES).text for res in restaurants]
+        assert names == ["Café Noir"]
+
+    def test_resource_id_unknown_id(self, client, restaurant, bad_restaurant):
+        """Prove that unknown IDs simply return an empty list."""
+        response = client.get(
+            "/v1/wfs/?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0&TYPENAMES=restaurant"
+            "&RESOURCEID=restaurant.0"
+        )
+        content = read_response(response)
+        assert response["content-type"] == "text/xml; charset=utf-8", content
+        assert response.status_code == 200, content
+        assert "</wfs:FeatureCollection>" in content
+
+        # Validate against the WFS 2.0 XSD
+        xml_doc = validate_xsd(content, WFS_20_XSD)
+        assert xml_doc.attrib["numberMatched"] == "0"
+        assert xml_doc.attrib["numberReturned"] == "0"
+
+        # Test sort ordering.
+        members = xml_doc.findall("wfs:member", namespaces=NAMESPACES)
+        assert len(members) == 0
+
+    def test_resource_id_typename_mismatch(self, client, restaurant, bad_restaurant):
+        """Prove that TYPENAMES should be omitted, or match the RESOURCEID."""
+        response = client.get(
+            "/v1/wfs/?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0"
+            "&TYPENAMES=mini-restaurant"
+            f"&RESOURCEID=restaurant.{restaurant.id}"
+        )
+        content = read_response(response)
+        assert response["content-type"] == "text/xml; charset=utf-8", content
+        assert response.status_code == 400, content
+        assert "</ows:Exception>" in content
+
+        xml_doc = validate_xsd(content, WFS_20_XSD)
+        assert xml_doc.attrib["version"] == "2.0.0"
+        exception = xml_doc.find("ows:Exception", NAMESPACES)
+        assert exception.attrib["exceptionCode"] == "InvalidParameterValue"
+
+        message = exception.find("ows:ExceptionText", NAMESPACES).text
+        assert message == (
+            "When TYPENAMES and RESOURCEID are combined, "
+            "the RESOURCEID type should be included in TYPENAMES."
+        )
+
+    def test_resource_id_invalid(self, client, restaurant, bad_restaurant):
+        """Prove that TYPENAMES should be omitted, or match the RESOURCEID."""
+        response = client.get(
+            "/v1/wfs/?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0&RESOURCEID=restaurant.ABC"
+        )
+        content = read_response(response)
+        assert response["content-type"] == "text/xml; charset=utf-8", content
+        assert response.status_code == 400, content
+        assert "</ows:Exception>" in content
+
+        xml_doc = validate_xsd(content, WFS_20_XSD)
+        assert xml_doc.attrib["version"] == "2.0.0"
+        exception = xml_doc.find("ows:Exception", NAMESPACES)
+        message = exception.find("ows:ExceptionText", NAMESPACES).text
+        assert exception.attrib["exceptionCode"] == "InvalidParameterValue", message
+        assert exception.attrib["locator"] == "resourceId", message
+        # message differs in Django versions
+
+    def test_resource_id_multiple(self, client, restaurant, bad_restaurant):
+        """Prove that fetching multiple IDs works."""
+        response = client.get(
+            "/v1/wfs/?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0"
+            f"&RESOURCEID=restaurant.{restaurant.id},restaurant.{bad_restaurant.id}"
+        )
+        content = read_response(response)
+        assert response["content-type"] == "text/xml; charset=utf-8", content
+        assert response.status_code == 200, content
+        assert "</wfs:FeatureCollection>" in content
+
+        # Validate against the WFS 2.0 XSD
+        xml_doc = validate_xsd(content, WFS_20_XSD)
+        assert xml_doc.attrib["numberMatched"] == "2"
+        assert xml_doc.attrib["numberReturned"] == "2"
+
+        # Test sort ordering.
+        restaurants = xml_doc.findall("wfs:member/app:restaurant", namespaces=NAMESPACES)
+        names = [res.find("app:name", namespaces=NAMESPACES).text for res in restaurants]
+        assert names == ["Café Noir", "Foo Bar"]
